requirements.txt
aiokafka>=0.10.0
structlog>=23.1.0
python-dotenv==1.0.0
